Some compat changes
Casted floats in size methods to int, strangle incompatibility. Currently building with py3.10, with some hacky modifications to aiojobs scheduler source to make it play nicely.

+++ galacteek/ui/browser/__init__.py
@@ -150,8 +150,8 @@ class JSConsoleWidget(QWidget):
         self.log('<p><b>Javascript output console</b></p>')
 
         self.setMinimumSize(
-            app.desktopGeometry.width() * 0.6,
-            app.desktopGeometry.height() * 0.8,
+            int(app.desktopGeometry.width() * 0.6),
+            int(app.desktopGeometry.height() * 0.8),
         )
 
     def onClose(self):
